# docs/swagger.yaml
definitions:
  api.Error:
    properties:
      error:
        type: string
    type: object
...
  rototiller.Storage:
    properties:
      create_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      customer_id:
        type: string
      id:
        type: string
      last_used:
        $ref: '#/definitions/timestamppb.Timestamp'
      name:
        type: string
      status:
        type: string
    type: object
  rototiller.Task:
    properties:
      kind:
        type: string
      params:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  timestamppb.Timestamp:
    properties:
      nanos:
        description: |-
          Non-negative fractions of a second at nanosecond resolution. Negative
          second values with fractions must still have non-negative nanos values
          that count forward in time. Must be from 0 to 999,999,999
          inclusive.
        type: integer
      seconds:
        description: |-
          Represents seconds of UTC time since Unix epoch
          1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
          9999-12-31T23:59:59Z inclusive.
        type: integer
    type: object
info:
  contact:
    email: logsquaredn@gmail.com
    name: logsquaredn
    url: https://rototiller.logsquaredn.io/
  license:
    name: logsquaredn
paths:
  /api.storage.v1.StorageService/CreateStorage:
    post:
      consumes:
      - application/json
      - ' application/zip'
      description: RPC Stores a dataset. The ID of this stored dataset can be used
        as input to jobs
      parameters:
      - description: Content type to be stored
        in: header
        name: X-Content-Type
        required: true
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      produces:
      - application/json
      responses:
        "2":
          description: ""
          schema:
            $ref: '#/definitions/api.Error'
        "5":
          description: ""
          schema:
            $ref: '#/definitions/api.Error'
        "16":
          description: ""
          schema:
            $ref: '#/definitions/api.Error'
      summary: RPC Create a storage
      tags:
      - Storage
  /api.storage.v1.StorageService/GetStorage:
    post:
      description: RPC Get the metadata of a stored dataset
      parameters:
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      produces:
      - application/json
      responses:
        "2":
          description: ""
          schema:
            $ref: '#/definitions/api.Error'
        "5":
          description: ""
          schema:
            $ref: '#/definitions/api.Error'
        "16":
          description: ""
          schema:
            $ref: '#/definitions/api.Error'
      summary: RPC Get a storage
      tags:
      - Storage
  /api.storage.v1.StorageService/GetStorageContent:
    post:
      description: RPC Gets the content of a stored dataset
      parameters:
      - description: Request results as a Zip or JSON. Default Zip
        in: header
        name: Accept
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      produces:
      - application/json
      - ' application/zip'
      responses:
        "2":
          description: ""
          schema:
            $ref: '#/definitions/api.Error'
        "16":
          description: ""
          schema:
            $ref: '#/definitions/api.Error'
      summary: RPC Get a storage's content
      tags:
      - Content
...
  /api/v1/storages:
    get:
      description: |-
        Get a list of stored datasets based on API Key

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Offset of storages to return
        in: query
        name: offset
        type: integer
      - description: Limit of storages to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rototiller.Storage'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a list of storage
      tags:
      - Storage
    post:
      consumes:
      - application/json
      - ' application/zip'
      description: |-
        Stores a dataset. The ID of this stored dataset can be used as input to jobs

        &emsp; - API Key is required either as a query parameter or a header
        &emsp; - Pass the geospatial data to be stored in the request body
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Storage name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rototiller.Storage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create a storage
      tags:
      - Storage
  /api/v1/storages/{id}:
    get:
      description: |-
        Get the metadata of a stored dataset

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Storage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rototiller.Storage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a storage
      tags:
      - Storage
  /api/v1/storages/{id}/content:
    get:
      description: |-
        Gets the content of a stored dataset

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: Request results as a Zip or JSON. Default Zip
        in: header
        name: Accept
        type: string
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Storage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - ' application/zip'
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a storage's content
      tags:
      - Content
  /api/v1/tasks:
    get:
      description: '&emsp; - API Key is required either as a query parameter or a
        header'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rototiller.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a list of task types
      tags:
      - Task
  /api/v1/tasks/{type}:
    get:
      description: '&emsp; - API Key is required either as a query parameter or a
        header'
      parameters:
      - description: Task type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rototiller.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a task type
      tags:
      - Task
swagger: "2.0"